revised file structure
