+++ dlib/media/ffmpeg_muxer.h
@@ -17,31 +17,83 @@ namespace dlib
 
         struct encoder_image_args
         {
-            int             h{0};
-            int             w{0};
-            AVPixelFormat   fmt{AV_PIX_FMT_YUV420P};
-            int             framerate{0};
+            /*!
+                WHAT THIS OBJECT REPRESENTS
+                    This class groups a set of arguments passed to the encoder and muxer classes.
+                    These must be set to non-zero or non-trivial values as they are used to configure 
+                    the underlying codec and optionally, an internal image scaler.
+                    Any frame that is pushed to encoder or muxer instances is resized to the codec's 
+                    pre-configured settings if their dimensions or pixel format don't match.
+                    For example, if the codec is configured to use height 512, width 384 and RGB format,
+                    using the variables below, and the frames already have these settings when pushed, 
+                    then no resizing is performed. If however they don't, then they are first resized. 
+            !*/
+
+            // Target height of codec.
+            int h{0};
+
+            // Target width of codec.
+            int w{0};
+            
+            // Target pixel format of codec.
+            AVPixelFormat fmt{AV_PIX_FMT_YUV420P};
+
+            // Target framerate of codec/muxer
+            int framerate{0};
         };
 
 // ---------------------------------------------------------------------------------------------------
 
         struct encoder_audio_args
         {
-            int             sample_rate{0};
-            uint64_t        channel_layout{AV_CH_LAYOUT_STEREO};
-            AVSampleFormat  fmt{AV_SAMPLE_FMT_S16};
+            /*!
+                WHAT THIS OBJECT REPRESENTS
+                    This class groups a set of arguments passed to the encoder and muxer classes.
+                    These must be set to non-zero or non-trivial values as they are used to configure 
+                    the underlying codec and optionally, an internal audio resampler.
+                    Any frame that is pushed to encoder or muxer instances is resampled to the codec's
+                    pre-configured settings if their sample format, sample rate or channel layout, don't match.
+            !*/
+
+            // Target sample rate of codec
+            int sample_rate{0};
+
+            // Target channel layout of codec
+            uint64_t channel_layout{AV_CH_LAYOUT_STEREO};
+
+            // Target sample format of codec
+            AVSampleFormat fmt{AV_SAMPLE_FMT_S16};
         };
 
 // ---------------------------------------------------------------------------------------------------
 
         struct encoder_codec_args
         {
-            AVCodecID                                    codec{AV_CODEC_ID_NONE};
-            std::string                                  codec_name;
+            /*!
+                WHAT THIS OBJECT REPRESENTS
+                    This class groups a set of arguments passed to the encoder and muxer classes.
+                    Some of these must be set to non-zero or non-trivial values as they are used 
+                    to configure the underlying codec. Others will only be used if non-zero or
+                    non-trivial.
+            !*/
+
+            // Codec ID used to configure the encoder. Either codec or codec_name MUST be set.
+            AVCodecID codec{AV_CODEC_ID_NONE};
+
+            // Codec name used to configure the encoder. This is used if codec == AV_CODEC_ID_NONE.
+            std::string codec_name;
+
+            // A dictionary of AVCodecContext and codec-private options. Used by "avcodec_open2()"
             std::unordered_map<std::string, std::string> codec_options;
-            int64_t                                      bitrate{-1};
-            int                                          gop_size{-1};
-            int                                          flags{0};
+
+            // Sets AVCodecContext::bit_rate if non-negative.
+            int64_t bitrate{-1};
+
+            // Sets AVCodecContext::gop_size if non-negative.
+            int gop_size{-1};
+
+            // OR-ed with AVCodecContext::flags if non-negative.
+            int flags{0};
         };
 
 // ---------------------------------------------------------------------------------------------------
@@ -49,42 +101,199 @@ namespace dlib
         class encoder
         {
         public:
+            /*!
+                WHAT THIS OBJECT REPRESENTS
+                    This class is a libavcodec wrapper which encodes video or audio to raw memory.
+                    Note, if you are creating a media file, it is easier to use the muxer object
+                    as it also works with raw codec files like .h264 files.
+                    This class is suitable for example if you need to send raw packets over a socket
+                    or interface with another library that requires encoded data, not raw images
+                    or raw audio samples.
+            !*/
+
             struct args
             {
+                /*!
+                    WHAT THIS OBJECT REPRESENTS
+                        This holds constructor arguments for encoder.
+                !*/
                 encoder_codec_args args_codec;
                 encoder_image_args args_image;
                 encoder_audio_args args_audio;
             };
 
-            encoder()                             = default;
-            encoder(encoder&& other)              = default;
-            encoder& operator=(encoder&& other)   = default;
+            encoder() = default;
+            /*!
+                ensures
+                    - is_open() == false
+            !*/
 
             encoder(
                 const args& a,
                 std::function<bool(std::size_t, const char*)> sink
             );
+            /*!
+                requires
+                    - a.args_codec.codec or a.args_codec.codec_name are set
+                    - Either a.args_image or a.args_audio is fully set
+                    - sink is set to a valid callback for writing packet data.
+                      dlib/media/sink.h contains callback wrappers for
+                      different buffer types.
+                ensures
+                    - Constructs encoder from args and sink
+                    - is_open() == true
+            !*/
 
+            encoder(encoder&& other) = default;
+            /*!
+                ensures
+                    - Move constructor
+                    - other is in an empty but otherwise valid state after move
+                    - other.is_open() == false after move
+            !*/
+
+            encoder& operator=(encoder&& other) = default;
+            /*!
+                ensures
+                    - Move assignment operator
+                    - other is in an empty but otherwise valid state after move
+                    - other.is_open() == false after move
+            !*/
+            
             ~encoder();
+            /*!
+                ensures
+                    - Destructor
+                    - flush() is called if it hasn't been already
+            !*/
 
-            bool            is_open()           const noexcept;
-            bool            is_image_encoder()  const noexcept;
-            bool            is_audio_encoder()  const noexcept;
-            AVCodecID       get_codec_id()      const noexcept;
-            std::string     get_codec_name()    const noexcept;
-            /*! video properties !*/
-            int             height()            const noexcept;
-            int             width()             const noexcept;
-            AVPixelFormat   pixel_fmt()         const noexcept;
-            int             fps()               const noexcept;
-            /*! audio properties !*/
-            int             sample_rate()       const noexcept;
-            uint64_t        channel_layout()    const noexcept;
-            AVSampleFormat  sample_fmt()        const noexcept;
-            int             nchannels()         const noexcept;
+            bool is_open() const noexcept;
+            /*!
+                ensures
+                    - Returns true if the codec is open and user may call push()
+            !*/
 
-            bool push(frame frame);
+            bool is_image_encoder() const noexcept;
+            /*!
+                ensures
+                    - Returns true if the codec is an image encoder.
+            !*/
+
+            bool is_audio_encoder() const noexcept;
+            /*!
+                ensures
+                    - Returns true if the codec is an audio encoder.
+            !*/
+
+            AVCodecID get_codec_id() const noexcept;
+            /*!
+                requires
+                    - is_open() == true
+                ensures
+                    - returns the codec id. See ffmpeg documentation or libavcodec/codec_id.h
+            !*/
+
+            std::string get_codec_name() const noexcept;
+            /*!
+                requires
+                    - is_open() == true
+                ensures
+                    - returns string representation of codec id.
+            !*/
+
+            int height() const noexcept;
+            /*!
+                requires
+                    - is_image_encoder() == true
+                ensures
+                    - returns the height of the configured codec, not necessarily the
+                      height of frames passed to push(frame)
+            !*/
+
+            int width() const noexcept;
+            /*!
+                requires
+                    - is_image_encoder() == true
+                ensures
+                    - returns the width of the configured codec, not necessarily the
+                      width of frames passed to push(frame)
+            !*/
+
+            AVPixelFormat pixel_fmt() const noexcept;
+            /*!
+                requires
+                    - is_image_encoder() == true
+                ensures
+                    - returns the pixel format of the configured codec, not necessarily the
+                      pixel format of frames passed to push(frame)
+            !*/
+
+            int fps() const noexcept;
+             /*!
+                requires
+                    - is_image_encoder() == true
+                ensures
+                    - returns the configured framerate of the codec.
+            !*/
+
+            int sample_rate() const noexcept;
+            /*!
+                requires
+                    - is_audio_encoder() == true
+                ensures
+                    - returns the sample rate of the configured codec, not necessarily the
+                      sample rate of frames passed to push(frame)
+            !*/
+
+            uint64_t channel_layout() const noexcept;
+            /*!
+                requires
+                    - is_audio_encoder() == true
+                ensures
+                    - returns the channel layout of the configured codec, not necessarily the
+                      channel layout of frames passed to push(frame).
+                      e.g. AV_CH_LAYOUT_STEREO, AV_CH_LAYOUT_MONO etc.
+            !*/
+
+            AVSampleFormat sample_fmt() const noexcept;
+            /*!
+                requires
+                    - is_audio_encoder() == true
+                ensures
+                    - returns the sample format of the configured codec, not necessarily the
+                      sample format of frames passed to push(frame)
+            !*/
+
+            int nchannels() const noexcept;
+            /*!
+                requires
+                    - is_audio_encoder() == true
+                ensures
+                    - returns the number of audio channels in the configured codec.
+            !*/
+
+            bool push(frame f);
+            /*!
+                requires
+                    - if is_image_encoder() == true, then f.is_image() == true
+                    - if is_audio_encoder() == true, then f.is_audio() == true
+                ensures
+                    - If f does not have matching settings to the codec, it is either
+                      resized or resampled before being pushed to the codec and encoded.
+                    - The callback passed to the constructor may or may not be invoked
+                      as the underlying resampler, audio fifo and codec may buffer.
+                    - Returns true if successfully encoded, even if callback wasn't invoked.
+                    - Returns false if either EOF, i.e. flush() has been previously called,
+                      or an error occurred, in which case is_open() == false.
+            !*/
+
             void flush();
+            /*!
+                ensures
+                    - Flushes the codec. Callback passed to constructor will likely be invoked.
+                    - is_open() == false
+                    - Becomes a no-op after the first time you call this.
+            !*/
 
         private:
             friend class muxer;
@@ -112,55 +321,287 @@ namespace dlib
         class muxer
         {
         public:
+            /*!
+                WHAT THIS OBJECT REPRESENTS
+                    This class is a libavformat wrapper which muxes video and/or audio streams to file.
+                    It is analogous to OpenCV's cv::VideoWriter class but is more configurable, supports audio, 
+                    devices (X11, speaker, ...) and network streams such as RTSP, HTTP, and more.
+                    Note that a video file, e.g. MP4, can contain multiple streams: video, audio, subtitles, etc.
+                    This class can encode both video and audio at the same time.
+            !*/
+
             struct args
             {
-                args() = default;
-                args(const std::string& filepath);
- 
-                std::string filepath;
-                std::string output_format;
-                std::unordered_map<std::string, std::string>  format_options;     //An AVDictionary filled with AVFormatContext and muxer-private options. Used by avformat_write_header()
-                std::unordered_map<std::string, std::string>  protocol_options;   //An AVDictionary filled with protocol-private options. Used by avio_open2()
+                /*!
+                    WHAT THIS OBJECT REPRESENTS
+                        This holds constructor arguments for muxer.
+                !*/
 
-                int max_delay{-1};  //See documentation for AVFormatContext::max_delay
-                std::chrono::milliseconds   connect_timeout{std::chrono::milliseconds::max()};
-                std::chrono::milliseconds   read_timeout{std::chrono::milliseconds::max()};
-                std::function<bool()>       interrupter;
+                args() = default;
+                /*!
+                    ensures
+                        - Default constructor.
+                !*/
+
+                args(const std::string& filepath);
+                /*!
+                    ensures
+                        - this->filepath = filepath
+                !*/
+
+                // Filepath, URL or device this object will write data to.
+                std::string filepath;
+
+                // Output format hint. e.g. 'rtsp', 'X11', 'alsa', etc. 99% of the time, users are not required to specify this as libavformat tries to guess it.
+                std::string output_format;
+
+                // A dictionary filled with AVFormatContext and muxer-private options. Used by "avformat_write_header()"".
+                // Please see libavformat documentation for more details
+                std::unordered_map<std::string, std::string> format_options;
+
+                // An AVDictionary filled with protocol-private options. Used by avio_open2()
+                std::unordered_map<std::string, std::string> protocol_options;
+
+                // See documentation for AVFormatContext::max_delay
+                int max_delay{-1};
+
+                // Only relevant to network muxers such as RTSP, HTTP etc.
+                // Connection/listening timeout. 
+                std::chrono::milliseconds connect_timeout{std::chrono::milliseconds::max()};
                 
+                // Only relevant to network muxers such as RTSP, HTTP etc
+                // Read timeout. 
+                std::chrono::milliseconds read_timeout{std::chrono::milliseconds::max()};
+
+                // Only relevant to network muxers such as RTSP, HTTP etc
+                // Connection/listening interruption callback.
+                // The constructor periodically calls interrupter() while waiting on the network. If it returns true, then
+                // the connection is aborted and the muxer is closed.
+                // So user may use this in conjunction with some thread-safe shared state to signal an abort/interrupt.
+                std::function<bool()> interrupter;
+
+                // Video stream arguments
                 struct : encoder_codec_args, encoder_image_args{} args_image;
+
+                // Audio stream arguments
                 struct : encoder_codec_args, encoder_audio_args{} args_audio;
+                
+                // Whether or not to encode video stream.
                 bool enable_image{true};
+
+                // Whether or not to encode audio stream.
                 bool enable_audio{true};
             };
 
             muxer() = default;
+            /*!
+                ensures
+                    - is_open() == false
+            !*/
+
             muxer(const args& a);
+            /*!
+                ensures
+                    - Constructs muxer using args
+            !*/
+
             muxer(muxer&& other) noexcept;
+            /*!
+                ensures
+                    - Move constructor
+                    - After move, other.is_open() == false
+            !*/
+
             muxer& operator=(muxer&& other) noexcept;
+            /*!
+                ensures
+                    - Move assignment
+                    - After move, other.is_open() == false
+            !*/
+
             ~muxer();
+            /*!
+                ensures
+                    - Calls flush() if not already called
+                    - Closes the underlying file/socket
+            !*/
 
             bool is_open() const noexcept;
+            /*!
+                ensures
+                    - Returns true if underlying container and codecs are open
+                    - User may call push()
+            !*/
+
             bool audio_enabled() const noexcept;
+            /*!
+                requires
+                    - args.enable_audio == true
+                ensures
+                    - returns true if is_open() == true and an audio stream is available and open
+            !*/
+
             bool video_enabled() const noexcept;
+            /*!
+                requires
+                    - args.enable_video == true
+                ensures
+                    - returns true if is_open() == true and a video stream is available and open
+            !*/
 
-            int             height()                    const noexcept;
-            int             width()                     const noexcept;
-            AVPixelFormat   pixel_fmt()                 const noexcept;
-            float           fps()                       const noexcept;
-            int             estimated_nframes()         const noexcept;
-            AVCodecID       get_video_codec_id()        const noexcept;
-            std::string     get_video_codec_name()      const noexcept;
+            int height() const noexcept;
+            /*!
+                ensures
+                    - if (video_enabled())
+                        - returns the height of the configured codec, not necessarily the height of
+                          frames passed to push(frame).
+                    - else
+                        - returns 0
+            !*/
 
-            int             sample_rate()               const noexcept;
-            uint64_t        channel_layout()            const noexcept;
-            AVSampleFormat  sample_fmt()                const noexcept;
-            int             nchannels()                 const noexcept;
-            int             estimated_total_samples()   const noexcept;
-            AVCodecID       get_audio_codec_id()        const noexcept;
-            std::string     get_audio_codec_name()      const noexcept;
+            int width() const noexcept;
+            /*!
+                ensures
+                    - if (video_enabled())
+                        - returns the width of the configured codec, not necessarily the
+                          width of frames passed to push(frame).
+                    - else
+                        - returns 0
+            !*/
+
+            AVPixelFormat pixel_fmt() const noexcept;
+            /*!
+                ensures
+                    - if (video_enabled())
+                        - returns the pixel format of the configured codec, not necessarily the
+                          pixel format of frames passed to push(frame).
+                    - else
+                        - returns AV_PIX_FMT_NONE
+            !*/
+
+            float fps() const noexcept;
+            /*!
+                ensures
+                    - if (video_enabled())
+                        - returns the framerate of the configured codec
+                    - else
+                        - returns 0
+            !*/
+
+            int estimated_nframes() const noexcept;
+            /*!
+                ensures
+                    - returns ffmpeg's estimation of how many video frames are in the file without reading any frames
+                    - See ffmpeg's documentation for AVStream::nb_frames
+                    - Note, this is known to not always work with ffmpeg v3 with certain files. Most of the time it does
+                      Do not make your code rely on this
+            !*/
+
+            AVCodecID get_video_codec_id() const noexcept;
+            /*!
+                ensures 
+                    - if (video_enabled())
+                        - returns codec ID of video stream
+                    - else
+                        - returns AV_CODEC_ID_NONE
+            !*/
+
+            std::string get_video_codec_name() const noexcept;
+            /*!
+                ensures 
+                    - if (video_enabled())
+                        - returns codec name of video stream
+                    - else
+                        - returns ""
+            !*/
+
+            int sample_rate() const noexcept;
+            /*!
+                ensures
+                    - if (audio_enabled())
+                        - returns the sample rate of the configured codec, not necessarily the
+                          sample rate of frames passed to push(frame).
+                    - else
+                        - returns 0
+            !*/
+
+            uint64_t channel_layout() const noexcept;
+            /*!
+                ensures
+                    - if (audio_enabled())
+                        - returns the channel layout of the configured codec, not necessarily the
+                          channel layout of frames passed to push(frame).
+                    - else
+                        - returns 0
+            !*/
+
+            int nchannels() const noexcept;
+            /*!
+                ensures
+                    - if (audio_enabled())
+                        - returns the number of audio channels in the configured codec, not necessarily the
+                          number of channels in frames passed to push(frame).
+                    - else
+                        - returns 0
+            !*/
+
+            AVSampleFormat sample_fmt() const noexcept;
+            /*!
+                ensures
+                    - if (audio_enabled())
+                        - returns the sample format of the configured codec, not necessarily the
+                          sample format of frames passed to push(frame).
+                    - else
+                        - returns AV_SAMPLE_FMT_NONE
+            !*/
+
+            int estimated_total_samples() const noexcept;
+            /*!
+                ensures
+                    - if (audio_enabled())
+                        - returns an estimation fo the total number of audio samples in the audio stream
+                    - else
+                        - returns 0
+            !*/
+
+            AVCodecID get_audio_codec_id() const noexcept;
+            /*!
+                ensures 
+                    - if (audio_enabled())
+                        - returns codec ID of audio stream
+                    - else
+                        - returns AV_CODEC_ID_NONE
+            !*/
+
+            std::string get_audio_codec_name() const noexcept;
+            /*!
+                ensures 
+                    - if (audio_enabled())
+                        - returns codec name of audio stream
+                    - else
+                        - returns ""
+            !*/
 
             bool push(frame f);
+            /*!
+                requires
+                    - if is_image_encoder() == true, then f.is_image() == true
+                    - if is_audio_encoder() == true, then f.is_audio() == true
+                ensures
+                    - If f does not have matching settings to the codec, it is either
+                      resized or resampled before being pushed to the muxer.
+                    - Encodes and writes the encoded data to file/socket
+                    - Returns true if successfully encoded.
+                    - Returns false if either EOF, i.e. flush() has been previously called,
+                      or an error occurred, in which case is_open() == false.
+            !*/
+
             void flush();
+            /*!
+                ensures
+                    - Flushes the file.
+                    - is_open() == false
+            !*/
 
         private:
